Network.AWS.LexRuntime.Types

Description

Service Configuration

lexRuntime :: Service

API version 2016-11-28 of the Amazon Lex Runtime Service SDK configuration.

Errors

_NotAcceptableException :: AsError a => Getting (First ServiceError) a ServiceError

The accept header in the request does not have a valid value.

_DependencyFailedException :: AsError a => Getting (First ServiceError) a ServiceError

One of the downstream dependencies, such as AWS Lambda or Amazon Polly, threw an exception. For example, if Amazon Lex does not have sufficient permissions to call a Lambda function, it results in Lambda throwing an exception.

_UnsupportedMediaTypeException :: AsError a => Getting (First ServiceError) a ServiceError

The Content-Type header (PostContent API) has an invalid value.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Two clients are using the same AWS account, Amazon Lex bot, and user ID.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource (such as the Amazon Lex bot or an alias) that is referred to is not found.

_RequestTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The input speech is too long.

_LoopDetectedException :: AsError a => Getting (First ServiceError) a ServiceError

Lambda fulfilment function returned DelegateDialogAction to Amazon Lex without changing any slot values.

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

Internal service error. Retry the call.

_BadGatewayException :: AsError a => Getting (First ServiceError) a ServiceError

Either the Amazon Lex bot is still building, or one of the dependent services (Amazon Polly, AWS Lambda) failed with an internal service error.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Request validation failed, there is no usable message in the context, or the bot build failed.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exceeded a limit.

ContentType

data ContentType

Constructors

ApplicationVnd_Amazonaws_Card_Generic

DialogState

data DialogState

Constructors

ConfirmIntent

ElicitIntent

ElicitSlot

Failed

Fulfilled

ReadyForFulfillment

Button

data Button

Represents an option to be shown on the client platform (Facebook, Slack, etc.)

See: button smart constructor.

button

Arguments

:: Text
bText

-> Text
bValue

-> Button

Creates a value of Button with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bText - Text that is visible to the user on the button.

bValue - The value sent to Amazon Lex when a user chooses the button. For example, consider button text "NYC." When the user chooses the button, the value sent can be "New York City."

bText :: Lens' Button Text

Text that is visible to the user on the button.

bValue :: Lens' Button Text

The value sent to Amazon Lex when a user chooses the button. For example, consider button text "NYC." When the user chooses the button, the value sent can be "New York City."

GenericAttachment

data GenericAttachment

Represents an option rendered to the user when a prompt is shown. It could be an image, a button, a link, or text.

See: genericAttachment smart constructor.

genericAttachment :: GenericAttachment

Creates a value of GenericAttachment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaButtons - The list of options to show to the user.

gaSubTitle - The subtitle shown below the title.

gaImageURL - The URL of an image that is displayed to the user.

gaAttachmentLinkURL - The URL of an attachment to the response card.

gaTitle - The title of the option.

gaButtons :: Lens' GenericAttachment [Button]

The list of options to show to the user.

gaSubTitle :: Lens' GenericAttachment (Maybe Text)

The subtitle shown below the title.

gaImageURL :: Lens' GenericAttachment (Maybe Text)

The URL of an image that is displayed to the user.

gaAttachmentLinkURL :: Lens' GenericAttachment (Maybe Text)

The URL of an attachment to the response card.

gaTitle :: Lens' GenericAttachment (Maybe Text)

The title of the option.

ResponseCard

data ResponseCard

If you configure a response card when creating your bots, Amazon Lex substitutes the session attributes and slot values that are available, and then returns it. The response card can also come from a Lambda function ( dialogCodeHook and fulfillmentActivity on an intent).

See: responseCard smart constructor.

responseCard :: ResponseCard

Creates a value of ResponseCard with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcGenericAttachments - An array of attachment objects representing options.

rcVersion - The version of the response card format.

rcContentType - The content type of the response.

rcGenericAttachments :: Lens' ResponseCard [GenericAttachment]

An array of attachment objects representing options.

rcVersion :: Lens' ResponseCard (Maybe Text)

The version of the response card format.

rcContentType :: Lens' ResponseCard (Maybe ContentType)

The content type of the response.